pretty much working code
node that the package name is changed

besides that User, userType, Report, Appointment, and DatabaseConnector
are all identical to their previous versions.

you have to have a local installation of XAMPP in order to emulate the
SQL server.  I will soon upload a backup of the SQL database that I
tested with.

I have already found 2 errors but I need to sleep and fix them later